feat: refactor message handling and add new file
- Remove `line_bot_api` import and initialization from the file `message_event.py`
- Add a new file `message_handler.py` with the necessary imports and functions
- Modify the `sendConfirm` function to handle cases where the `result[6]` value is empty or contains specific strings
- Modify the `sendQuickreply` function to include different text prompts based on the `typeButton` value
- Modify the `sendContent` function to handle cases where the `result[5]` value is empty or contains specific strings
- Modify the `sendList` function to include additional text messages in the response
